Move MqttClient to own module
This patch moves the `MqttClient` type to a `client` module.

For this, some of the functions on `MqttClient` had to be made
`pub(crate)` and some `pub(crate)`-getters had to be introduced.
